requirements.txt
Flask
flask-mysqldb
PyYAML
scikit-learn
pandas